requirements.txt
# DEAN Orchestration Requirements

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Async HTTP
httpx==0.25.2
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development tools
black==23.11.0
ruff==0.1.6
mypy==1.7.1
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Utilities
pyyaml==6.0.1
click==8.1.7
rich==13.7.0

# WebSocket support
websockets==12.0

# Service mesh
tenacity==8.2.3

# Apache Airflow client
apache-airflow-client==2.7.0

# Evolution and ML (optional)
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2